Added parent Manager class: Code refactoring and reusing
